Count Characters in String
Simple function to iterate through a string and log the number of times
each unique character occurs.  I chose to exercise using a reducer on a
character array generated from the passed string.

Also included are a few primitive tests, and a "deep" comparison for
simple objects of 1 layer (with no other nested objects).  I don't
really check if properties exist out-of-order in each object.
Properties must exist in the same order with the same value in each
object being compared.