Memoized Fibonacci
This set of functions "memoizes" Fibonacci so that the stack won't
overflow.  One is simply a cache and return cached value vanilla
variant; the other is an iterative Fibonacci that scans from beginning
(F[0] and F[1]) to user value (F[n]), summing the previous two
associative array entries for a speed boost and less overhead.

These operate on BigInts due to the inherent "bigness" (largeness?)
of Fibonacci values - the cache is seeded with BigInt[0] and BigInt[1],
and the rest is keeping our recursion / iteration values clean.