Brainfuck Interpreter
This is a really, really crude Brainfuck interpreter.  It has really
terrible integer based bracket matching, and scans everytime a
block/loop ([]) is activated.  There is no caching or optimization or
anything neat here.

I plan to get around to doing basic refactoring and cleanup and maybe
some "neatifying" of this.