Find Three Most Frequent Words
This set of functions tokenizes and scans an input text for "words",
keeping a count for duplicates.  An array of the 3 most frequent words
is returned, sorted descending by frequency.  The findNMostFrequent()
function should scale to return however many frequent words you want to
find - if there are none, an empty array is returned.

Includes JSDoc'ing, basic console / terminal assertions and testing, and
a few other things.